persisted_stm: removes stm_snapshot_recovery_policy
The way we write snapshot files (a write to a temporary file and then
atomic move operation) prevents us from a common case leading to broken
file (incomplete write caused by an app crash).

So it an over-engineering to assume it's going be such a common problem
that it needs a custom solution. Removing this functionality to make
persisted_stm simpler.

If customers run into this problem they can simply remove the snapshot.
It will have the same consistency and availability impact.