Update URL for cert generation script (#586)
To make this script usable for the Console `docker-compose.yml` example as well as `localhost`, I asked Juan to update his script so that it generates alternative CNs for:

- localhost
- redpanda
- console
- connect
- 127.0.0.1

This PR modifies the docs page to reflect the new URL for the Gist. 

We should consider moving that script out of Juan's account and into an official Redpanda location.